Moved Ranking and RankingId to ranking entity folder
