+++ src/main/java/tennisclub/dao/RankingDaoImpl.java
@@ -1,8 +1,7 @@
 package tennisclub.dao;
 
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-import tennisclub.entity.Ranking;
+import tennisclub.entity.ranking.Ranking;
 import tennisclub.entity.Tournament;
 import tennisclub.entity.User;
