+++ src/main/java/tennisclub/entity/ranking/RankingId.java
@@ -1,4 +1,7 @@
-package tennisclub.entity;
+package tennisclub.entity.ranking;
+
+import tennisclub.entity.Tournament;
+import tennisclub.entity.User;
 
 import java.io.Serializable;
